Edit establishConnection method so it only asks for a url
The establishConnection method currently takes a url, user, and password
(all separately). However, in my attempts to build a JDBC url that will
work with Google Cloud, the only method I've found that might work is a
url format that has the user and password built into it. Therefore, we
want to prompt for only a url with everything in it.
 # Please enter the commit message for your changes. Lines starting